src/test/java/org/water/chapter1/CustomerServiceTest.java
package org.water.chapter1;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.water.chapter1.model.Customer;
import org.water.chapter1.service.CustomerService;

import java.io.BufferedReader;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

/**
 * Created by 帝 on 2017/1/1.
 */
public class CustomerServiceTest {
    private final CustomerService customerService;
    public CustomerServiceTest(){
        customerService = new CustomerService();
    }

    @Before
    public void init()throws Exception{
        String file = "sql/customer_init.sql";
        DBHelper.executeSqlFile(file);
    }

    @Test
    public void getCustomerListTest() throws Exception{
        List<Customer> customerList = customerService.getCustomerList(null);
        Assert.assertEquals(2, customerList.size());
    }

    @Test
    public void getCustomerTest() throws Exception{
        long id = 1;
        Customer customer = customerService.getCustomer(id);
        //System.out.println(customer.getName() + customer.getContact());
        Assert.assertNotNull(customer);

    }

    @Test
    public void createCustomerTest() throws Exception{
        Map<String, Object> fieldMap = new HashMap<String, Object>();
        //fieldMap.put("id", 1); error
        fieldMap.put("name", "TestName");
        fieldMap.put("contact", "Test");
        fieldMap.put("telephone", "123");
        boolean result = customerService.createCustomer(fieldMap);
        Assert.assertTrue(result);
    }

    @Test
    public void updateCustomerTest() throws Exception{
        long id = 1;
        Map<String, Object> fieldMap = new HashMap<String, Object>();
        fieldMap.put("contact", "Eric");
        boolean result = customerService.updateCustomer(id, fieldMap);
        Assert.assertTrue(result);
    }

    @Test
    public void deleteCustomerTest() throws Exception{
        long id = 2;
        boolean result = customerService.deleteCustomer(id);
        Assert.assertTrue(result);
    }
}
